+++ R/duckdb.R
@@ -26,14 +26,16 @@ duckdb_connection <- function(..., extensions = character(0), add_auth = TRUE) {
 #' @param ... Unused
 #' @param conn A DuckDB connection
 #' @param read_func The DuckDB SQL function to call against the list of urls.
-#'                  Defaults to `read_parquet`.
+#'                  One of `"read_parquet"` (default) or `"read_csv"`.
 #' @export
 tbl_http <- function(
   urls,
   ...,
   conn = duckdb_connection(extensions = "httpfs"),
-  read_func = c("read_parquet", "read_csv", )
+  read_func = c("read_parquet", "read_csv")
 ) {
+
+  read_func <- match.arg(read_func)
 
   # TODO(justin): allow read_func arguments i.e. union_by_name
   query <- paste0("SELECT * FROM ", read_func, "([",